Always run via subprocess in python-selector.
If we run via import then we don't execute any of the code
that checks for 'if __name__ == "__main__"' which means that
target script needs to be aware its running under the selector.

Also, since we support both 2 and 3 for now we might as well
always run under `sys.executable` which we know for sure exists
(unlike `python2` which doesn't always exist on mac for example).